Varios errores corregidos

+++ src/main/java/backend/siptis/service/editorsAndReviewers/ProjectTribunalServiceImpl.java
@@ -1,5 +1,6 @@
 package backend.siptis.service.editorsAndReviewers;
 
+import backend.siptis.auth.entity.SiptisUser;
 import backend.siptis.commons.Phase;
 import backend.siptis.commons.ServiceMessage;
 import backend.siptis.commons.ServiceAnswer;
@@ -7,6 +8,7 @@ import backend.siptis.model.entity.editorsAndReviewers.ProjectTribunal;
 import backend.siptis.model.entity.projectManagement.Defense;
 import backend.siptis.model.entity.projectManagement.Project;
 import backend.siptis.model.pjo.dto.editorsAndReviewers.ReviewADefenseDTO;
+import backend.siptis.model.pjo.dto.projectManagement.AssignTribunalsDTO;
 import backend.siptis.model.pjo.vo.projectManagement.ProjectToTribunalHomePageVO;
 import backend.siptis.model.repository.editorsAndReviewers.ProjectTribunalRepository;
 import backend.siptis.model.repository.projectManagement.ProjectRepository;
@@ -15,6 +17,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -112,7 +115,7 @@ public class ProjectTribunalServiceImpl implements ProjectTribunalService {
 
         Project project = query.getProject();
 
-        if(!project.getPhase().equals(Phase.TRIBUNALS_PHASE.toString())){
+        if(!project.getPhase().equals(Phase.ASSIGN_TRIBUNALS_PHASE.toString())){
             return ServiceAnswer.builder().serviceMessage(ServiceMessage.PROJECT_IS_ON_ANOTHER_PHASE).data(null).build();
         }
 
@@ -179,6 +182,10 @@ public class ProjectTribunalServiceImpl implements ProjectTribunalService {
             return ServiceAnswer.builder().serviceMessage(ServiceMessage.TRIBUNALS_ERROR).data("PROJECT DOES NOT HAVE TRIBUNALS TO REMOVE").build();
         }
 
+        if(project.getTotalDefensePoints() != null){
+            return ServiceAnswer.builder().serviceMessage(ServiceMessage.TRIBUNALS_ERROR).data("THIS PROJECT HAS BEEN DEFENDED").build();
+        }
+
         for (ProjectTribunal tribunal: project.getTribunals()) {
             projectTribunalRepository.delete(tribunal);
         }
@@ -209,5 +216,36 @@ public class ProjectTribunalServiceImpl implements ProjectTribunalService {
                 .collect(Collectors.toList());
 
         return ServiceAnswer.builder().serviceMessage(ServiceMessage.OK).data(data).build();
+    }
+
+    @Override
+    public ServiceAnswer assignTribunals(AssignTribunalsDTO assignTribunalsDTO) {
+        List<Long> tribunalsIds = assignTribunalsDTO.getTribunalsIds();
+        Long projectId = assignTribunalsDTO.getProjectId();
+        List<SiptisUser> tribunals = new ArrayList<>();
+        for(Long id: tribunalsIds){
+            Optional<SiptisUser> user = siptisUserRepository.findById(id);
+            if(user.isEmpty()){
+                return ServiceAnswer.builder().serviceMessage(ServiceMessage.USER_ID_DOES_NOT_EXIST).data(id).build();
+            }
+            if(user.get().getRoles().stream().noneMatch(role -> role.getName().equals("TRIBUNAL"))){
+                return ServiceAnswer.builder().serviceMessage(ServiceMessage.USER_IS_NOT_A_TRIBUNAL).data(id).build();
+            }
+            tribunals.add(user.get());
+        }
+        Optional<Project> project = projectRepository.findById(projectId);
+        if(project.isEmpty()){
+            return ServiceAnswer.builder().serviceMessage(ServiceMessage.PROJECT_ID_DOES_NOT_EXIST).data(null).build();
+        }
+
+        for(SiptisUser user: tribunals){
+            ProjectTribunal projectTribunal = new ProjectTribunal(user, project.get());
+            projectTribunalRepository.save(projectTribunal);
+        }
+
+        Project projectNotOptional = project.get();
+        projectNotOptional.setPhase(Phase.PRE_DEFENSE_PHASE.toString());
+
+        return ServiceAnswer.builder().serviceMessage(ServiceMessage.OK).data("Tribunals has been assigned to the project").build();
     }
 }
